LPS-43746 Remove builders since they don't serve any purpose
We can remove these helper classes. They were intended to serve as a fluent API to create Sharepoint XML documents but they will be costly to maintain and the code is not much cleaner than when using the constructors directly. Also, as we have moved away from MS's nomenclature, the fluent API is not useful any more.